pyproject.toml
[tool.poetry]
name = "froods"
version = "0.1.0"
description = ""
authors = ["Frank Rosner"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.40.2"
google-auth = "^2.33.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
google-api-python-client = "^2.140.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
